task_discord/task_discord/requirements.txt
discord.py
sqlite3